prevent the photo modal dialogs from overflowing the page
The photo upload dialog was fine, but change width to min-width so we
don't squash the footer & buttons. It doesn't too wide because of an
inherited rule "width: 560px" from bootstrap.

Copy the styling of the photo upload dialogs to the badge cropping:
modal container 70% height
modal body 80% of that
nested <img> 100% of its available space.

Tested with small & large browser window. Large photos.